Add new Django app - PeopleTracker
Added folder structure by language - the new app is in Python-Projects. This might be
interesting for Nik B's senior project. Its purpose is to determine how full the hallways
may be in a transition. Will need a lot of work instantiating models (maybe a script can
be hooked up in the school's database?).

	* Added folder structure
	* Added PeopleTracker
	* Updated README