Convert PeopleTracker to Heroku repo
I'm seeing if this "nested repository" thing will work. In PeopleTracker, there is
a Heroku repository that I can push to a heroku app (https://asdstudent.herokuapp.com/).
Lots of additions to that. See asdstudent.herokuapp.com/admin to add new students to
improve the functionality of the PeopleTracker app.